Ensure that fixtures created with identical methods are different.
`self.thing` was being stored in the test case's `__dict__` to check
if the fixture had been created before.  Since multiple fixtures can
be created, calling the same method, this didn't work.  Only the first
fixture created for `self.thing` would be used.

Instead, store the application of the fixture (calling the method with
the args and kwargs) in the tet case's `__dict__`, keyed by the fixture
itself.

Fixes #14.